フィールド名やスペルミスの修正
