+++ src/main/java/StudentManagement/data/Student.java
@@ -1,4 +1,4 @@
-package StudentManagement;
+package StudentManagement.data;
 
 
 import lombok.Getter;
