+++ src/main/java/StudentManagement/data/StudentCourse.java
@@ -1,4 +1,4 @@
-package StudentManagement;
+package StudentManagement.data;
 
 
 import java.time.LocalDate;
