+++ src/main/java/StudentManagement/repository/StudentCourseRepository.java
@@ -1,5 +1,6 @@
-package StudentManagement;
+package StudentManagement.repository;
 
+import StudentManagement.data.StudentCourse;
 import java.util.List;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Select;
